Debug de parte de los warnings y limpieza de archivos
