+++ data_model/dimension_enum_dc.py
@@ -1,6 +1,4 @@
-import enum
-from dataclasses import dataclass
-from data_model.ine_status import Status
 from collections import namedtuple
 
-DimensionItem = namedtuple("DimensionItem",["id","name", "columns", "dataframe","values_df","status_df", "data", "values", "status"])
+DimensionItem = namedtuple("DimensionItem",  ["id", "name", "columns", "dataframe", "values_df", "status_df", "data",
+                                              "values", "status"])
